Franchise-elite/requirements.txt
Django==2.2.6
django-filter==2.2.0
django-tables2==2.1.1
Pillow==6.2.1
pip==19.0.3
pytz==2019.3
setuptools==40.8.0
sqlparse==0.3.0
